Tidied up factorial
The now-omitted line was left over from editing the previous version. But it occurred to me that the loop takes care of of that case already, so I removed the extraneous line. It doesn't do anything.

I also removed the spacing to look more like the scrunched up style of the rest. ;p